Bug fixes on resolving method and member types as well as handling nested classes.
Some fixes to stateChartExample.xml.  One problem is the use of members in non-static functions. When the members are wrapped as Parameter<?>s, they need to be unwrapped with a getValue().  Maybe it just needs a simple search/replace on all members.

Made edits to the generated Scenario_latest.xml (which was copied from Scenario_XML_2012-242T15.07.38).  The receive effect needs attention.