Added horizon duration to XML schema (and to many of the checked in XML scenarios).  Domains of Timepoint and Duration are by default restricted to the horizon.  Activities outside the horizon will not elaborate.  This avoid infinite event generation for perpetual activity (like a customer always using power).
Created a "latest" package to hold the generated files from Scenario_latest.xml that we hand edited.  We should now delete files in generated.

Some cleanup of the test code so that it can find the right XML files.  Added a FileUtils.findFile() function to accomplish this since trying to get the files as resources either fails or finds sometimes old versions of them in the bin directory.