Created a MagicDrawExecutor to tie the Java simulation to the Python MagidDrawAnimator.  Not tested.
Developed AE.py to integrate everything to run together.  It needs to hook into a plugin to test the MagicDrawExecutor.

EventXmlToJava.java has a lot of functions for runtime compile, load, and run of generated classes.  This class needs to be broken up--it does too many things and is getting bloated.