Parsing Class.expr to avoid errors parsing things like "Power_System.Signalyes.class."
Attempted to also provide support for expressions like "min(a,b)" which should be "Math.min(a,b)," but the code is not yet able to match them.

This is using org.apache.commons.lang, which could break the build for others, in which case it should be replaced with some commented out code.