fixed an infinite loop of pickValue() by not calling the ParameterConstraint.Helper.pickValue() from ConstraintExpression.  This is correct since if pickValue fails on this path, the Parameter will pick it's own, just as the Helper would have.
The nestedCall in Call was nesting in the wrong direction.  Now the result of the nested call's evaluate() is used as the object whose method is invoked.  Before the result of this.evaluate() would be used as the object for the nestedCall's object, and the result of the nested call's evaluate() would be passed back as the result of this.evaluate().  This didn't match the Java code being generated from XML.