Replaced trees with hash tables for performance improvement.  Now that we generate our own hash codes, hash tables should be safe.
More improved performance was gained by passing false in for deep for some get???() functions.  The expectation that deep gets are unnecessary in some instances where events are already being gathered deep.  The Smaller.Customer simulation now generates roughly the same simulations with no constraint violations, but this may be caused by incomplete gathering of constraints.

One idea is to traverse shallow most of the time but deep when progress (increased # of satisfied constraints) is not being made.

Some cosmetic changes to the text output was made to ignore certain expected errors and provide more information about intermediate progress.